requirements.txt
pandas==2.2.2
requests==2.31.0
pyarrow==15.0.2
pytest==8.2.0
apache-airflow==2.8.1